Rename test to spec
