+++ src/client/authenticator_spec.rs
