+++ src/client/storage_spec.rs
